Fix windows crate 0.58 API breakages and CRT mismatch in karukan-tsf
Adapt all TSF COM interface implementations to windows 0.58 breaking changes:
- ITfKeyEventSink: remove pfeaten out-param, return Result<BOOL>, add OnSetFocus
- Registry APIs: append .ok()? for WIN32_ERROR, fix RegisterProfile signature
- IClassFactory/DllGetClassObject: handle query() returning HRESULT directly
- RequestEditSession: remove &mut hr out-param, use Result<HRESULT> return
- RegisterGUID: use Result<u32> return instead of out-param
- AsImpl::as_impl: returns value directly instead of Result
- Lang bar: remove .0 on style constants, fix Self:: resolution in _Impl
- Wrap unsafe calls in explicit unsafe blocks (Rust 2024 unsafe-in-unsafe-fn)
- Fix InsertTextAtSelection flag type

Also fix CRT mismatch by forcing CMAKE_MSVC_RUNTIME_LIBRARY=MultiThreaded
in .cargo/config.toml, and resolve pre-existing clippy warnings.

+++ karukan-tsf/src/keymap.rs
@@ -69,9 +69,9 @@ pub fn vk_to_keysym(vk: u32, unicode_char: Option<char>) -> Keysym {
 
         // Alphanumeric fallback (when unicode_char is not available)
         // VK codes for 'A'-'Z' are 0x41-0x5A, map to lowercase keysym
-        vk @ 0x41..=0x5A => Keysym((vk - 0x41 + 0x61) as u32), // 'a'-'z'
+        vk @ 0x41..=0x5A => Keysym(vk - 0x41 + 0x61), // 'a'-'z'
         // VK codes for '0'-'9' are 0x30-0x39
-        vk @ 0x30..=0x39 => Keysym(vk as u32),
+        vk @ 0x30..=0x39 => Keysym(vk),
 
         // IME-specific keys
         0x15 => Keysym(0xff2a), // VK_KANJI / Hankaku/Zenkaku — mapped to Kanji keysym
